+++ src/core_modules/capture-core/components/Pages/EnrollmentAddEvent/NewEventWorkspace/newEventWorkspace.types.js
@@ -12,5 +12,9 @@ export type Props = {|
     rulesExecutionDependencies: Object,
     onSave: ExternalSaveHandler,
     onCancel: () => void,
+|};
+
+export type PlainProps = {|
+    ...Props,
     ...CssClasses
 |};