+++ src/core_modules/capture-core/components/WidgetProfile/DataEntry/FormFoundation/DataElement.js
@@ -19,14 +19,22 @@ import { convertFormToClient, convertClientToServer } from '../../../../converte
 import { convertOptionSetValue } from '../../../../converters/serverToClient';
 import { buildIcon } from '../../../../metaDataMemoryStoreBuilders/common/helpers';
 import { OptionGroup } from '../../../../metaData/OptionSet/OptionGroup';
-import { getFeatureType, getDataElement, getLabel, isNotValidOptionSet, escapeString } from '../helpers';
+import {
+    getFeatureType,
+    getDataElement,
+    getLabel,
+    escapeString,
+    handleAPIResponse,
+    REQUESTED_ENTITIES,
+} from '../helpers';
+import {
+    handleUnsupportedMultiText,
+} from '../../../../metaDataMemoryStoreBuilders/common/helpers/dataElement/unsupportedMultiText';
 import type { QuerySingleResource } from '../../../../utils/api/api.types';
 
 const OPTION_SET_NOT_FOUND = 'Optionset not found';
 const TRACKED_ENTITY_ATTRIBUTE_NOT_FOUND =
     'TrackedEntityAttributeId missing from programTrackedEntityAttribute or trackedEntityAttribute not found';
-const MULIT_TEXT_WITH_NO_OPTIONS_SET =
-    'could not create the metadata because a MULIT_TEXT without associated option sets was found';
 
 const buildDataElementUnique = (
     dataElement: DataElement,
@@ -81,7 +89,8 @@ const buildDataElementUnique = (
                 });
             }
             return requestPromise.then((result) => {
-                const otherTrackedEntityInstances = result?.instances?.filter(item => item.trackedEntity !== contextProps.trackedEntityInstanceId) || [];
+                const apiTrackedEntities = handleAPIResponse(REQUESTED_ENTITIES.trackedEntities, result);
+                const otherTrackedEntityInstances = apiTrackedEntities.filter(item => item.trackedEntity !== contextProps.trackedEntityInstanceId);
                 const trackedEntityInstance = (otherTrackedEntityInstances && otherTrackedEntityInstances[0]) || {};
 
                 const data = {
@@ -145,6 +154,7 @@ const buildBaseDataElement = async (
     programTrackedEntityAttribute: ProgramTrackedEntityAttribute,
     trackedEntityAttribute: TrackedEntityAttribute,
     querySingleResource: QuerySingleResource,
+    minorServerVersion: number,
 ) => {
     const dataElement = new DataElement();
     dataElement.type = trackedEntityAttribute.valueType;
@@ -155,11 +165,8 @@ const buildBaseDataElement = async (
         trackedEntityAttribute,
         querySingleResource,
     });
-    if (isNotValidOptionSet(dataElement.type, dataElement.optionSet)) {
-        log.error(errorCreator(MULIT_TEXT_WITH_NO_OPTIONS_SET)({ dataElement }));
-        return null;
-    }
-    return dataElement;
+
+    return handleUnsupportedMultiText(dataElement, minorServerVersion);
 };
 
 const buildDateDataElement = async (
@@ -249,6 +256,7 @@ export const buildDataElement = (
     trackedEntityAttributes: Array<TrackedEntityAttribute>,
     optionSets: Array<OptionSetType>,
     querySingleResource: QuerySingleResource,
+    minorServerVersion: number,
 ) => {
     const trackedEntityAttribute =
         programTrackedEntityAttribute.trackedEntityAttributeId &&
@@ -268,5 +276,5 @@ export const buildDataElement = (
 
     return trackedEntityAttribute.valueType === dataElementTypes.DATE
         ? buildDateDataElement(optionSets, programTrackedEntityAttribute, trackedEntityAttribute, querySingleResource)
-        : buildBaseDataElement(optionSets, programTrackedEntityAttribute, trackedEntityAttribute, querySingleResource);
+        : buildBaseDataElement(optionSets, programTrackedEntityAttribute, trackedEntityAttribute, querySingleResource, minorServerVersion);
 };