+++ src/core_modules/capture-core/metaDataMemoryStoreBuilders/common/helpers/dataElement/unsupportedMultiText/isMultiTextWithoutOptionset.js
@@ -1,0 +1,7 @@
+// @flow
+import { dataElementTypes } from '../../../../../metaData';
+
+export const isMultiTextWithoutOptionset = (
+    valueType: string,
+    optionSet?: ?Object,
+) => valueType === dataElementTypes.MULTI_TEXT && !optionSet;