+++ src/core_modules/capture-core/metaDataMemoryStoreBuilders/programs/factory/enrollment/DataElementFactory.js
@@ -1,6 +1,7 @@
 // @flow
 /* eslint-disable no-underscore-dangle */
 import log from 'loglevel';
+import { handleAPIResponse, REQUESTED_ENTITIES } from 'capture-core/utils/api';
 import i18n from '@dhis2/d2-i18n';
 import { pipe, errorCreator } from 'capture-core-utils';
 
@@ -21,7 +22,9 @@ import { OptionSetFactory } from '../../../common/factory';
 import { convertFormToClient, convertClientToServer } from '../../../../converters';
 import type { ConstructorInput } from './dataElementFactory.types';
 import type { QuerySingleResource } from '../../../../utils/api/api.types';
-import { isNotValidOptionSet } from '../../../../utils/isNotValidOptionSet';
+import {
+    handleUnsupportedMultiText,
+} from '../../../common/helpers/dataElement/unsupportedMultiText';
 import { escapeString } from '../../../../utils/escapeString';
 
 export class DataElementFactory {
@@ -106,7 +109,8 @@ export class DataElementFactory {
                 }
                 return requestPromise
                     .then((result) => {
-                        const otherTrackedEntityInstances = result?.instances?.filter(item => item.trackedEntity !== contextProps.trackedEntityInstanceId) || [];
+                        const apiTrackedEntities = handleAPIResponse(REQUESTED_ENTITIES.trackedEntities, result);
+                        const otherTrackedEntityInstances = apiTrackedEntities.filter(item => item.trackedEntity !== contextProps.trackedEntityInstanceId);
                         const trackedEntityInstance = (otherTrackedEntityInstances && otherTrackedEntityInstances[0]) || {};
 
                         const data = {
@@ -130,10 +134,12 @@ export class DataElementFactory {
     locale: ?string;
     optionSetFactory: OptionSetFactory;
     cachedTrackedEntityAttributes: Map<string, CachedTrackedEntityAttribute>;
+    minorServerVersion: number;
     constructor({
         cachedTrackedEntityAttributes,
         cachedOptionSets,
         locale,
+        minorServerVersion,
     }: ConstructorInput) {
         this.cachedTrackedEntityAttributes = cachedTrackedEntityAttributes;
         this.locale = locale;
@@ -141,6 +147,7 @@ export class DataElementFactory {
             cachedOptionSets,
             locale,
         );
+        this.minorServerVersion = minorServerVersion;
     }
 
     _getAttributeTranslation(
@@ -219,11 +226,7 @@ export class DataElementFactory {
             cachedProgramTrackedEntityAttribute,
             cachedTrackedEntityAttribute,
         );
-        if (isNotValidOptionSet(dataElement.type, dataElement.optionSet)) {
-            log.error(errorCreator(DataElementFactory.errorMessages.MULIT_TEXT_WITH_NO_OPTIONS_SET)({ dataElement }));
-            return null;
-        }
-        return dataElement;
+        return handleUnsupportedMultiText(dataElement, this.minorServerVersion);
     }
 
     async _buildDateDataElement(