+++ src/core_modules/capture-core/metaDataMemoryStoreBuilders/trackedEntityTypes/factory/TrackedEntityType/DataElementFactory.js
@@ -1,6 +1,7 @@
 // @flow
 /* eslint-disable no-underscore-dangle */
 import log from 'loglevel';
+import { handleAPIResponse, REQUESTED_ENTITIES } from 'capture-core/utils/api';
 import i18n from '@dhis2/d2-i18n';
 import { pipe, errorCreator } from 'capture-core-utils';
 import type {
@@ -18,8 +19,10 @@ import { OptionSetFactory } from '../../../common/factory';
 import { convertFormToClient, convertClientToServer } from '../../../../converters';
 import type { ConstructorInput } from './dataElementFactory.types';
 import type { QuerySingleResource } from '../../../../utils/api/api.types';
-import { isNotValidOptionSet } from '../../../../utils/isNotValidOptionSet';
 import { escapeString } from '../../../../utils/escapeString';
+import {
+    handleUnsupportedMultiText,
+} from '../../../../metaDataMemoryStoreBuilders/common/helpers/dataElement/unsupportedMultiText';
 
 export class DataElementFactory {
     static translationPropertyNames = {
@@ -51,10 +54,12 @@ export class DataElementFactory {
     locale: ?string;
     optionSetFactory: OptionSetFactory;
     cachedTrackedEntityAttributes: Map<string, CachedTrackedEntityAttribute>;
+    minorServerVersion: number;
     constructor({
         cachedTrackedEntityAttributes,
         cachedOptionSets,
         locale,
+        minorServerVersion,
     }: ConstructorInput) {
         this.cachedTrackedEntityAttributes = cachedTrackedEntityAttributes;
         this.locale = locale;
@@ -62,6 +67,7 @@ export class DataElementFactory {
             cachedOptionSets,
             locale,
         );
+        this.minorServerVersion = minorServerVersion;
     }
 
     _getAttributeTranslation(
@@ -77,7 +83,7 @@ export class DataElementFactory {
 
     async build(
         cachedTrackedEntityTypeAttribute: CachedTrackedEntityTypeAttribute,
-    ) {
+    ): Promise<DataElement | null> {
         const cachedAttribute = cachedTrackedEntityTypeAttribute.trackedEntityAttributeId &&
             this.cachedTrackedEntityAttributes.get(
                 cachedTrackedEntityTypeAttribute.trackedEntityAttributeId,
@@ -168,15 +174,15 @@ export class DataElementFactory {
                     }
                     return requestPromise
                         .then((result) => {
-                            const trackedEntityInstance =
-                                (result.instances && result.instances[0]) || {};
+                            const apiTrackedEntities = handleAPIResponse(REQUESTED_ENTITIES.trackedEntities, result);
+                            const trackedEntityInstance = apiTrackedEntities[0] || {};
                             const data = {
                                 id: trackedEntityInstance.trackedEntity,
                                 tetId: trackedEntityInstance.trackedEntityType,
                             };
 
                             return {
-                                valid: result.instances.length === 0,
+                                valid: apiTrackedEntities.length === 0,
                                 data,
                             };
                         });
@@ -196,14 +202,8 @@ export class DataElementFactory {
                 null,
                 value => value,
             );
-            if (isNotValidOptionSet(dataElement.type, dataElement.optionSet)) {
-                log.error(
-                    errorCreator(DataElementFactory.errorMessages.MULIT_TEXT_WITH_NO_OPTIONS_SET)({ dataElement }),
-                );
-                return null;
-            }
         }
 
-        return dataElement;
+        return handleUnsupportedMultiText(dataElement, this.minorServerVersion);
     }
 }