+++ src/core_modules/capture-core/metaDataMemoryStoreBuilders/trackedEntityTypes/factory/TrackedEntityType/dataElementFactory.types.js
@@ -8,4 +8,5 @@ export type ConstructorInput = {|
     cachedTrackedEntityAttributes: Map<string, CachedTrackedEntityAttribute>,
     cachedOptionSets: Map<string, CachedOptionSet>,
     locale: ?string,
+    minorServerVersion: number,
 |};