+++ src/core_modules/capture-core/metaDataMemoryStoreBuilders/trackedEntityTypes/trackedEntityTypesBuilder.types.js
@@ -10,4 +10,5 @@ export type BuildTrackedEntityTypesInput = {|
     cachedTrackedEntityAttributes: Map<string, CachedTrackedEntityAttribute>,
     cachedOptionSets: Map<string, CachedOptionSet>,
     locale: ?string,
+    minorServerVersion: number,
 |};