+++ src/core_modules/capture-core/reducers/descriptions/enrollmentDomain.reducerDescription.js
@@ -3,12 +3,9 @@ import { createReducerDescription } from '../../trackerRedux';
 import { enrollmentSiteActionTypes } from '../../components/Pages/common/EnrollmentOverviewDomain';
 import { actionTypes as enrollmentNoteActionTypes }
     from '../../components/WidgetEnrollmentComment/WidgetEnrollmentComment.actions';
-import {
-    actionTypes as editEventActionTypes,
-} from '../../components/WidgetEventEdit/EditEventDataEntry/editEventDataEntry.actions';
-import {
-    newEventWidgetActionTypes,
-} from '../../components/WidgetEnrollmentEventNew/Validated/validated.actions';
+import { actionTypes as editEventActionTypes } from '../../components/WidgetEventEdit/EditEventDataEntry/editEventDataEntry.actions';
+import { newEventWidgetActionTypes } from '../../components/WidgetEnrollmentEventNew/Validated/validated.actions';
+import { enrollmentEditEventActionTypes } from '../../components/Pages/EnrollmentEditEvent';
 
 const initialReducerValue = {};
 const {
@@ -24,6 +21,17 @@ const {
     COMMIT_ENROLLMENT_EVENTS,
     ADD_PERSISTED_ENROLLMENT_EVENTS,
 } = enrollmentSiteActionTypes;
+
+const setAssignee = (state, action) => {
+    const { assignedUser, eventId } = action.payload;
+
+    const events = state.enrollment.events.reduce(
+        (acc, e) => (e.event === eventId ? [...acc, { ...e, assignedUser }] : [...acc, e]),
+        [],
+    );
+
+    return { ...state, enrollment: { ...state.enrollment, events } };
+};
 
 export const enrollmentDomainDesc = createReducerDescription(
     {
@@ -168,6 +176,8 @@ export const enrollmentDomainDesc = createReducerDescription(
             const { eventSaveInProgress, ...newState } = state;
             return newState;
         },
+        [enrollmentEditEventActionTypes.ASSIGNEE_SET]: setAssignee,
+        [enrollmentEditEventActionTypes.ASSIGNEE_SAVE_FAILED]: setAssignee,
     },
     'enrollmentDomain',
     initialReducerValue,