+++ src/core_modules/capture-core/utils/isNotValidOptionSet.js
@@ -1,9 +1,0 @@
-// @flow
-import { Option, dataElementTypes } from '../metaData';
-
-export const isNotValidOptionSet = (
-    valueType: string,
-    optionSet?: ?{
-        +options: Array<Option>,
-    },
-) => valueType === dataElementTypes.MULTI_TEXT && (!optionSet || optionSet.options.length === 0);